merge scl only, binarization formula and ard step

+++ run_s2c.py
@@ -1,7 +1,8 @@
 import json
 
 import click
-from dataship.dag.utils import l2a_to_ard
+from dataship.dag.utils import l2a_to_ard, binary_scl
+from dataship.dag.s2_dag import get_s2_product
 from ewoc_db.fill.update_status import get_next_tile
 
 from utils import *
@@ -74,32 +75,47 @@ def run_plan(plan, l2a_dir, provider, config):
 @click.option("-p", "--pid", help="S2 L1C product ID")
 @click.option("-o", "--l2a_dir", default=None, help="Output directory")
 @click.option("-cfg", "--config", default=None, help="EOdag config file")
-@click.option("-pv", "--provider", default="creodias", help="Data provider")
-def run_id(pid, l2a_dir, provider, config):
+@click.option("-sc", "--only_scl", default=False, is_flag=True)
+@click.option("-pv", "--provider", default="creodias_eodata", help="Data provider")
+@click.option('--force_push', is_flag=True)
+@click.option('--no_sen2cor', help="Do not process with Sen2cor", is_flag=True)
+def run_id(pid, l2a_dir, provider, config, only_scl=False, force_push=False, no_sen2cor=False):
     if l2a_dir is None:
         l2a_dir = "/work/SEN2TEST/OUT/"
-    # Generate temporary folders
-    dem_tmp_dir = "/work/SEN2TEST/DEM/"
-    tile = pid.split("_")[5][1:]
-    if not os.path.exists(dem_tmp_dir):
-        os.makedirs(dem_tmp_dir)
-    dem_syms = custom_s2c_dem(tile, tmp_dir=dem_tmp_dir)
-    out_dir_l1c, out_dir_l2a = make_tmp_dirs(l2a_dir)
-    # Get Sat product by id using eodag
-    robust_get_by_id(pid, out_dir_l1c)
-    l1c_safe_folder = [
-        os.path.join(out_dir_l1c, fold) for fold in os.listdir(out_dir_l1c) if fold.endswith("SAFE")
-    ][0]
-    l1c_safe_folder = last_safe(l1c_safe_folder)
-    # Run sen2cor in subprocess
-    l2a_safe_folder = run_s2c(l1c_safe_folder, out_dir_l2a)
-    # Convert the sen2cor output to ewoc ard format
-    l2a_to_ard(l2a_safe_folder, l2a_dir)
-    # Delete local folders
-    clean(out_dir_l2a)
-    clean(out_dir_l1c)
-    clean(dem_tmp_dir)
-    unlink(dem_syms)
+        if not os.path.exists(l2a_dir):
+            os.makedirs(l2a_dir)
+    if "L2A" in pid and not no_sen2cor:
+        raise AttributeError("Using L2A product with Sen2cor is impossible")
+    if no_sen2cor:
+        if only_scl:
+            get_s2_product(pid, l2a_dir, source=provider)
+            scl_to_ard(l2a_dir, pid)
+        else:
+            raise NotImplementedError("Only the SCL MASK production is implemented")
+    else:
+        # Generate temporary folders
+        dem_tmp_dir = "/work/SEN2TEST/DEM/"
+        tile = pid.split("_")[5][1:]
+        if not os.path.exists(dem_tmp_dir):
+            os.makedirs(dem_tmp_dir)
+        dem_syms = custom_s2c_dem(tile, tmp_dir=dem_tmp_dir)
+
+        out_dir_l1c, out_dir_l2a = make_tmp_dirs(l2a_dir)
+        # Get Sat product by id using eodag
+        robust_get_by_id(pid, out_dir_l1c, provider)
+        l1c_safe_folder = [
+            os.path.join(out_dir_l1c, fold) for fold in os.listdir(out_dir_l1c) if fold.endswith("SAFE")
+        ][0]
+        l1c_safe_folder = last_safe(l1c_safe_folder)
+        # Run sen2cor in subprocess
+        l2a_safe_folder = run_s2c(l1c_safe_folder, out_dir_l2a, only_scl)
+        # Convert the sen2cor output to ewoc ard format
+        l2a_to_ard(l2a_safe_folder, l2a_dir, only_scl)
+        # Delete local folders
+        clean(out_dir_l2a)
+        clean(out_dir_l1c)
+        clean(dem_tmp_dir)
+        unlink(dem_syms)
     # Send to s3
     ewoc_s3_upload(l2a_dir)
